% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipfun.R
\name{fdn}
\alias{fdn}
\title{Find name in dataframe}
\usage{
fdn(data, name)
}
\arguments{
\item{data}{- the data object to look into}

\item{name}{- the text string to search for}
}
\description{
Return the matched (ignore case) names in the data obj.
}
\examples{
fdn(mtcars, 'p') # Return all the column names containing 'p'
}
\keyword{find}
\keyword{name}
