% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipfun.R
\name{ipr_fixname}
\alias{ipr_fixname}
\title{Return fixed name of input dataframe}
\usage{
ipr_fixname(df, rmlist = c(" ", "(", ")", ":", "_", "-", "\\\\", "/"))
}
\arguments{
\item{df}{- the input dataframe}

\item{rmlist}{- special chars to remove, default: " ", "(", ")", ":", "_", "-", "\\", "/"}
}
\description{
Fix the input dataframe names by removing special characters such as "_"
}
\examples{
names(dataframe_tobe_fixed) = ipr_fixname(dataframe_tobe_fixed)
}
\keyword{fix}
\keyword{name}
