% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipfun.R
\name{ldn}
\alias{ldn}
\title{Find name in global env}
\usage{
ldn(name)
}
\arguments{
\item{name}{- the text string to search for}
}
\description{
Return the matched (ignore case) names in the global env.
}
\examples{
ldn('d') # Return all the column names containing 'd'
}
\keyword{find}
\keyword{name}
