% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipfun.R
\name{lsfun}
\alias{lsfun}
\title{List all available functions in a package}
\usage{
lsfun(package_name, output.list = F)
}
\arguments{
\item{name}{- package_name: input package to list its function}
}
\description{
Return all the functions in the input: package_name
}
\examples{
lsfun("dplyr") # Return all the functions in dplyr package (tree view)
lsfun('dplyr', output.list=T) # Return all the functions as list
}
\keyword{function}
\keyword{list}
\keyword{pacakage}
