% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipfun.R
\name{lu}
\alias{lu}
\title{Return the length of unique items}
\usage{
lu(obj)
}
\arguments{
\item{obj}{- the object/list to find the length of unique items}
}
\description{
Return the length of unique items for the given argument
}
\examples{
lu(df$keyid) # Return the length of unique items in df$keyid
}
\keyword{length}
\keyword{unique}
