% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipfun.R
\name{mkdirp}
\alias{mkdirp}
\title{Create folder similarly as mkdir -p}
\usage{
mkdirp(folder)
}
\arguments{
\item{folder}{- the path of the folder to be created}
}
\description{
Create folder similarly as mkdir -p
}
\examples{
mkdirp("./folder_to_create")
}
